Protein Features pipeline - partial annotation and xrefs

+++ modules/Bio/EnsEMBL/Production/Pipeline/ProteinFeatures/InterProScan.pm
@@ -57,60 +57,65 @@ sub run {
   my $outfile_xml  = "$outfile_base.xml";
   my $outfile_tsv  = "$outfile_base.tsv";
   
-  if ($run_interproscan) {
-    unlink $outfile_xml if -e $outfile_xml;
-    unlink $outfile_tsv if -e $outfile_tsv;
-    
-    # Must use /scratch for short temporary file names; TMHMM can't cope with longer ones.
-    my $scratch_dir = tempdir(DIR => '/scratch', CLEANUP => 1);
-    
-    if (! -e $scratch_dir) {
-      $self->warning("Output directory '$scratch_dir' does not exist. I shall create it.");
-      make_path($scratch_dir) or $self->throw("Failed to create output directory '$scratch_dir'");
+  if (scalar(@$applications)) {
+    if ($run_interproscan) {
+      unlink $outfile_xml if -e $outfile_xml;
+      unlink $outfile_tsv if -e $outfile_tsv;
+      
+      # Must use /scratch for short temporary file names; TMHMM can't cope with longer ones.
+      my $scratch_dir = tempdir(DIR => '/scratch', CLEANUP => 1);
+      
+      if (! -e $scratch_dir) {
+        $self->warning("Output directory '$scratch_dir' does not exist. I shall create it.");
+        make_path($scratch_dir) or $self->throw("Failed to create output directory '$scratch_dir'");
+      }
+      
+      my $options = "--iprlookup --goterms --pathways ";
+      $options .= "-f TSV, XML -t $seq_type --tempdir $scratch_dir ";
+      $options .= '--applications '.join(',', @$applications).' ';
+      my $input_option  = "-i $input_file ";
+      my $output_option = "--output-file-base $outfile_base ";  
+      
+      if ($run_mode =~ /^(nolookup|local)$/) {
+        $options .= "--disable-precalc ";
+      }
+      
+      my $interpro_cmd = qq($interproscan_exe $options $input_option $output_option);
+      
+      if ($self->param_is_defined('species')) {
+        my $dba = $self->get_DBAdaptor('core');
+        $dba->dbc && $dba->dbc->disconnect_if_idle();
+      }
+      $self->dbc and $self->dbc->disconnect_if_idle();
+      
+      $self->run_cmd($interpro_cmd);
+      
+      unlink $scratch_dir if -e $scratch_dir;
     }
     
-    my $options = "--iprlookup --goterms --pathways ";
-    $options .= "-f TSV, XML -t $seq_type --tempdir $scratch_dir ";
-    $options .= '--applications '.join(',', @$applications).' ';
-    my $input_option  = "-i $input_file ";
-    my $output_option = "--output-file-base $outfile_base ";  
-    
-    if ($run_mode =~ /^(nolookup|local)$/) {
-      $options .= "--disable-precalc ";
+    if (! -e $outfile_xml) {
+      $self->throw("Output file '$outfile_xml' was not created");
+    } elsif (-s $outfile_xml == 0) {
+      $self->throw("Output file '$outfile_xml' was zero size");
     }
     
-    my $interpro_cmd = qq($interproscan_exe $options $input_option $output_option);
-    
-    if ($self->param_is_defined('species')) {
-      my $dba = $self->get_DBAdaptor('core');
-      $dba->dbc && $dba->dbc->disconnect_if_idle();
-    }
-    $self->dbc and $self->dbc->disconnect_if_idle();
-    
-    system($interpro_cmd) == 0 or $self->throw("Failed to run ".$interpro_cmd);
-    
-    unlink $scratch_dir if -e $scratch_dir;
+    $self->param('outfile_xml', $outfile_xml);
+    $self->param('outfile_tsv', $outfile_tsv);
   }
-  
-  if (! -e $outfile_xml) {
-    $self->throw("Output file '$outfile_xml' was not created");
-  } elsif (-s $outfile_xml == 0) {
-    $self->throw("Output file '$outfile_xml' was zero size");
-  }
-  
-  $self->param('outfile_xml', $outfile_xml);
-  $self->param('outfile_tsv', $outfile_tsv);
 }
 
 sub write_output {
   my ($self) = @_;
+  my $applications = $self->param_required('interproscan_applications');
   
-  my $output_ids =
-  {
-    'outfile_xml' => $self->param('outfile_xml'),
-    'outfile_tsv' => $self->param('outfile_tsv'),
-  };
-  $self->dataflow_output_id($output_ids, 1);
+  if (scalar(@$applications)) {
+    my $output_ids =
+    {
+      'outfile_xml' => $self->param('outfile_xml'),
+      'outfile_tsv' => $self->param('outfile_tsv'),
+    };
+    $self->dataflow_output_id($output_ids, 3);
+  }
 }
 
 1;
